refactor(VMenu) convert to typescript (#7098)
* refactor(VMenu): merge mixins into main menu

* refactor(VMenu): change extension to ts

* refactor(VMenu): convert functionality to ts
